+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Change Log
 * Adicionar código do veículo no retorno das buscas de serviço (PL-3524)
 * Remove placa de todo o fluxo de aferições (PL-3403)
 * Adiciona código de veículo no cronograma de aferição (PL-3496)
-* Remove placa na realização de checklist (PL-3545)
+* Remove placa das buscas do checklist (PL-3546)
 
 <a name="v2.0.3"></a>
 ## Version [v2.0.3](https://github.com/luizfp/PrologWebService/compare/v2.0.2...v2.0.3) (release-date) [unreleased]